+++ code/src/cljs/sixsq/nuvla/ui/session/views.cljs
@@ -142,8 +142,11 @@
       [ui/DropdownItem {:aria-label (@tr [:support])
                         :icon       "mail"
                         :text       (@tr [:support])
-                        :href       (str "mailto:support%40sixsq%2Ecom?subject=%5BSlipStream%5D%20"
-                                         "Support%20question%20%2D%20Not%20logged%20in")}]]]))
+                        :href       (js/encodeURI
+                                      (str "mailto:support@sixsq.com?subject=[Nuvla] "
+                                           "Support question - "
+                                           (if logged-in? @user "Not logged in")))
+                        }]]]))
 
 
 (defn authn-menu