+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+  - Authn menu - Fix support text and set user when logged in
   - Edge details - allow add and edit of location on Map
   - Edge - NuvlaBox displayed on map
   - Map leaflet integration